use polymake_run_script instead of directly running polymake
