+++ deps/build_Perl.v5.30.3.jl
@@ -9,15 +9,15 @@ products = [
 ]
 
 # Download binaries from hosted location
-bin_prefix = "https://github.com/benlorenz/perl_jll.jl/releases/download/Perl-v5.30.3+0"
+bin_prefix = "https://github.com/JuliaBinaryWrappers/Perl_jll.jl/releases/download/Perl-v5.30.3+0"
 
 # Listing of files generated by BinaryBuilder:
 download_info = Dict(
-    Linux(:i686, libc=:glibc) => ("$bin_prefix/Perl.v5.30.3.i686-linux-gnu.tar.gz", "ade7e9d0c67dfb9fb6123d19847ea264d63455a77ad79883c75d0dc08e2c20eb"),
-    Linux(:i686, libc=:musl) => ("$bin_prefix/Perl.v5.30.3.i686-linux-musl.tar.gz", "c86e671ff5869cb3913a86d7aaaf121c51d4569491006041df61af03577a7278"),
-    MacOS(:x86_64) => ("$bin_prefix/Perl.v5.30.3.x86_64-apple-darwin14.tar.gz", "1e066b76ef3053009c34d9e8c3743dac4f7830c28722bc999b294a265fba63c6"),
-    Linux(:x86_64, libc=:glibc) => ("$bin_prefix/Perl.v5.30.3.x86_64-linux-gnu.tar.gz", "3de938cfa5b4d25b9fa764517f59577eb6691b8380abaf39626d8445fc5846b0"),
-    Linux(:x86_64, libc=:musl) => ("$bin_prefix/Perl.v5.30.3.x86_64-linux-musl.tar.gz", "627dc5e836531945f4fca0f7e7312da14ad458b9ed11b692b02d6c100c80e6ed"),
+    Linux(:i686, libc=:glibc) => ("$bin_prefix/Perl.v5.30.3.i686-linux-gnu.tar.gz", "9f43143efaf0583bae902f579d64d25e3f5157a3ba2bf153540effba5abaabb4"),
+    Linux(:i686, libc=:musl) => ("$bin_prefix/Perl.v5.30.3.i686-linux-musl.tar.gz", "3884f2c7a373e169c94a6c604f484f54d71f97091826e7861b9d07dee99dd50f"),
+    MacOS(:x86_64) => ("$bin_prefix/Perl.v5.30.3.x86_64-apple-darwin14.tar.gz", "123a6bb9dc7bc8889184dfa2342f78eb61ffad188c9f997469a1ca2cab2025c0"),
+    Linux(:x86_64, libc=:glibc) => ("$bin_prefix/Perl.v5.30.3.x86_64-linux-gnu.tar.gz", "0413f7ca7e11fb8bfb6c3d68d2afb8b43d1bdf3b174187da794af93a1da1aab5"),
+    Linux(:x86_64, libc=:musl) => ("$bin_prefix/Perl.v5.30.3.x86_64-linux-musl.tar.gz", "546a3bc32f34c76ed8211c3bb0659d0dc0410a936fbc515d876f48ef6964eda5"),
 )
 
 # Install unsatisfied or updated dependencies:
