+++ deps/build_libpolymake_julia.v0.0.2.jl
@@ -1,24 +1,22 @@
 using BinaryProvider # requires BinaryProvider 0.3.0 or later
+import Libdl.dlext
 
 # Parse some basic command-line arguments
 const verbose = "--verbose" in ARGS
 const prefix = Prefix(get([a for a in ARGS if a != "--verbose"], 1, joinpath(@__DIR__, "usr")))
-
-products = Product[
-    LibraryProduct(prefix, "libpolymake", :libpolymake)
-    ExecutableProduct(prefix,"polymake", :polymake)
-    ExecutableProduct(prefix,"polymake-config", Symbol("polymake_config"))
+products = [
+    ExecutableProduct(prefix, "polymake_run_script", :polymake_run_script),
+    FileProduct(prefix, joinpath("lib","libpolymake_julia.$(dlext)"), :libpolymake_julia),
+    FileProduct(prefix, joinpath("share","libpolymake_julia","type_translator.jl"), :type_translator)
 ]
 
 # Download binaries from hosted location
-bin_prefix = "https://github.com/benlorenz/polymake_jll.jl/releases/download/polymake-v4.1.0+0"
+bin_prefix = "https://github.com/benlorenz/libpolymake_julia_jll.jl/releases/download/libpolymake_julia-v0.0.2+0"
 
 # Listing of files generated by BinaryBuilder:
 download_info = Dict(
-    MacOS(:x86_64, compiler_abi=CompilerABI(:gcc_any, :cxx03)) => ("$bin_prefix/polymake.v4.1.0.x86_64-apple-darwin14-cxx03.tar.gz", "5f57b69067cf618b65a841c5bdf05a47983989557abec3386fbdb590ac15642b"),
-    Linux(:x86_64, libc=:glibc, compiler_abi=CompilerABI(:gcc6, :cxx11)) => ("$bin_prefix/polymake.v4.1.0.x86_64-linux-gnu-gcc6-cxx11.tar.gz", "ccb4a44242a20e3b377fb46d0933d0d696e8f4b0657610ba155e97a3d48a2b25"),
-    Linux(:x86_64, libc=:glibc, compiler_abi=CompilerABI(:gcc7, :cxx11)) => ("$bin_prefix/polymake.v4.1.0.x86_64-linux-gnu-gcc7-cxx11.tar.gz", "aa2c294c1166d047fec55414d47cd423d8f0040c789dfff726b329e4ecb6cc9d"),
-    Linux(:x86_64, libc=:glibc, compiler_abi=CompilerABI(:gcc8, :cxx11)) => ("$bin_prefix/polymake.v4.1.0.x86_64-linux-gnu-gcc8-cxx11.tar.gz", "132f2e9455fadb333b392d584b51722e3fc2d8a1fe733c43a36a98662f8ff97f"),
+    MacOS(:x86_64, compiler_abi=CompilerABI(:gcc_any, :cxx11)) => ("$bin_prefix/libpolymake_julia.v0.0.2.x86_64-apple-darwin14-cxx11.tar.gz", "5a6626c2d03a2cb8833cc7bf03753bc7e0536fd7b6bf793adc31e7b8d620123f"),
+    Linux(:x86_64, libc=:glibc, compiler_abi=CompilerABI(:gcc_any, :cxx11)) => ("$bin_prefix/libpolymake_julia.v0.0.2.x86_64-linux-gnu-cxx11.tar.gz", "28e70589f95703d64d072d920e5b2fe5fb69261d019ee550585964768192641b"),
 )
 
 # Install unsatisfied or updated dependencies:
